// src/types/resident.ts
export interface ResidentDto {
  id: string | null;
  userId: string;
  apartmentBuildingId: string;
  name: string;
  brithDay: string | null;
  identityNumber: string | null;
  memberType: string;
  phoneNumber: string | null;
  email: string;
  userName: string;
  password: string;
  apartmentId: string;
  isOwner: boolean;
}